Adding double support did not fix numerical precision errors.
Former-commit-id: 1588677b0fd7feb772c00968b8ca4a19733c5397

+++ Designs/Untitled.py
@@ -8,7 +8,7 @@ add_preprocessor_define(define="""
 #define union(a,b) T_min(a,b)
 #define intersection(a,b) T_max(a,b)
 #define subtraction(a,b) T_max(a,-b)
-#define Vector3f(x,y,z) ((float3)((float)(x),(float)(y),(float)(z)))
+#define Vector3f(x,y,z) ((double3)((float)(x),(float)(y),(float)(z)))
 #define signOfInt(i) (i>0?1:(i<0?-1:(0)))
 
 #define DIRECTION_X 0
@@ -87,12 +87,12 @@ define_auxillary_function(function="""
 	}
 
 
-	float maxComponent(float3 v){
+	float maxComponent(double3 v){
 
 		return T_max(v.x,T_max(v.y,v.z));
 	}
 
-	float box(float3 point, float3 center, float3 halfDiameter ){
+	float box(double3 point, double3 center, double3 halfDiameter ){
 
 		point=fabs(point-center);
 
@@ -100,7 +100,7 @@ define_auxillary_function(function="""
 		
 	}
 
-	float getComponent(float3 v, int component){
+	float getComponent(double3 v, int component){
 
 		if(component==0) return v.x;
 		if(component==1) return v.y;
@@ -110,19 +110,19 @@ define_auxillary_function(function="""
 	
 	}
 
-	float3 termProduct(float3 a, float3 b){
+	double3 termProduct(double3 a, double3 b){
 
 		return Vector3f(a.x*b.x,a.y*b.y,a.z*b.z);
 	}
 
-	float3 swizzle(float3 v, int a, int b, int c){
+	double3 swizzle(double3 v, int a, int b, int c){
 		return Vector3f(getComponent(v,a),getComponent(v,b),getComponent(v,c));
 	}
 
 
 
 	//Messed up the orientation at first
-	float _hilbertUnitCell(float3 v){
+	float _hilbertUnitCell(double3 v){
 
 		float d1 = box(v,Vector3f(-0.5,-0.5,0.0),Vector3f(lineWidth,lineWidth,0.5+lineWidth));
 		float d2 = box(v,Vector3f(0.5,-0.5,0.0),Vector3f(lineWidth,lineWidth,0.5+lineWidth));
@@ -151,7 +151,7 @@ define_auxillary_function(function="""
 	}
 
 	//Messed up the orientation at first
-	float hilbertUnitCell(float3 v){
+	float hilbertUnitCell(double3 v){
 
 		v=termProduct(swizzle(v,1,0,2),Vector3f(1,-1,1));
 		v=termProduct(swizzle(v,2,1,0),Vector3f(1,1,-1));
@@ -160,9 +160,9 @@ define_auxillary_function(function="""
 	}
 
 
-	float putHilbert(float3 v,int x, int y, int z)
+	float putHilbert(double3 v,int x, int y, int z)
 	{
-		float3 c = Vector3f(x/3.0,y/3.0,z/3.0);
+		double3 c = Vector3f(x/3.0,y/3.0,z/3.0);
 		v=Vector3f(v.x-c.x,v.y-c.y,v.z-c.z);
 		v=Vector3f(3.0*v.x,3.0*v.y,3.0*v.z);
 
@@ -183,9 +183,9 @@ define_auxillary_function(function="""
 		float m21=quadrantMatrices[matrixOffset+7];
 		float m22=quadrantMatrices[matrixOffset+8];
 
-		float3 mc0 = Vector3f(m00,m01,m02);
-		float3 mc1 = Vector3f(m10,m11,m12);
-		float3 mc2 = Vector3f(m20,m21,m22); 
+		double3 mc0 = Vector3f(m00,m01,m02);
+		double3 mc1 = Vector3f(m10,m11,m12);
+		double3 mc2 = Vector3f(m20,m21,m22); 
 
 		float A = dot(v,mc0);
 		float B = dot(v,mc1);
@@ -195,7 +195,7 @@ define_auxillary_function(function="""
 
 	}
 
-	float putShaft(float3 v, float3 center, float halfWidth, float halfLength, int direction){
+	float putShaft(double3 v, double3 center, float halfWidth, float halfLength, int direction){
 		float d = MAX_DISTANCE;
 		switch(direction){
 			case DIRECTION_X:
@@ -215,9 +215,9 @@ define_auxillary_function(function="""
 		return d;
 	}
 
-	float putConnector(float3 v, int largeI, int largeJ, int largeK, int i, int j, int k, int direction){
+	float putConnector(double3 v, int largeI, int largeJ, int largeK, int i, int j, int k, int direction){
 
-		float3 center = Vector3f(
+		double3 center = Vector3f(
 
 (largeI*1.0+i/2.0)*1/3.0,
 (largeJ*1.0+j/2.0)*1/3.0,
@@ -229,7 +229,7 @@ define_auxillary_function(function="""
 
 	}
 
-	float putConnectors(float3 v){
+	float putConnectors(double3 v){
 
 		float d = MAX_DISTANCE;
 		d=union(d,putConnector(v,0,-1,1,0,1,1,DIRECTION_X));
@@ -248,7 +248,7 @@ define_auxillary_function(function="""
 	}
 
 
-	float hilbert_sdf(float3 v){
+	float hilbert_sdf(double3 v){
 
 
 		//return hilbertUnitCell(v);
